pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neo4j-asset-manager"
version = "0.1.0"
description = "Neo4j Asset Manager using Google Cloud Vertex AI"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Kevin"},
]
dependencies = [
    "google-genai>=1.18.0",
    "google-cloud-aiplatform>=1.44.0",
    "google-api-python-client>=2.122.0",
    "google-cloud-core>=2.4.1",
    "google-cloud-storage>=2.7.0",
    "neo4j>=5.13.0",
    "graphdatascience",
    "retry==0.9.2",
    "pydantic>=2.5.0",
    "streamlit>=1.27.2",
    "streamlit-chat>=0.1.1",
    "streamlit-chat-media>=0.0.4",
    "numpy",
    "pandas",
    "requests",
    "plotly>=5.20.0",
    "tqdm",
    "pillow",
    "altair<5",
    "python-dotenv",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.0.286",
]


[tool.hatch.build.targets.wheel]
packages = ["src/neo4j_asset_manager"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
